requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
mysql-connector-python==8.3.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.4
eventlet==0.35.1